+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { RouterLink , RouterOutlet } from '@angular/router';
-import { MessageService } from './message.service';
+import { MessageService } from './services/message.service';
 
 @Component ({
   selector: 'app-root' ,