+++ src/app/services/person.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Person } from './person';
+import { Person } from '../models/person';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';